Move dirtyRange logic to GlyphTexture
- it's only used there and complicates our simple
  Texture unnecessarily
